Different output file for every task
Previously every task was using the same name as output file, so that if
you execute multiple tasks during the same build, every task was
overwriting the previous output.

This commit fix this issue providing a different name per each task. The
name is generated using the same source set name that is used for the
task name. This is also aligned to the output of Android Lint.